% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{calculation of areas}
\usage{
area(h, w, fig)
}
\arguments{
\item{h}{height or radius of the figure.}

\item{w}{weight of the figure.}

\item{fig}{type of geometrical figure, "rectangle" (r), "triangle" (t), "circle" (c)}
}
\value{
A float.
}
\description{
calculation of areas
}
\examples{
hg <- 3
wg <- 2
fig <-"triangle"
area(hg, wg, fig)
}
